Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
eventregistry